モジュール構成を変更、 tests を追加
